Clarification about options and flags in the docs. (#87)
The documentation did not make 100% explicit the difference between options and flags, also, seems like it used the term 'flag' in the place of the term 'option' by the end of the text. The '-like' in flag-like and non-flag-like was also not helping (why flag-like? it IS a flag by the definition of 'flag' that is provided, and the context was not related to parsing, so it was not discussing if it some token appeared to be a flag or not to the parser).